added routes, survey, homepage, and working on friends page
