+++ app/data/friends.js
@@ -1,0 +1,82 @@
+var friends = [
+    {
+        name: "Ein",
+        photo: "https://www.google.com/url?sa=i&source=images&cd=&ved=&url=https%3A%2F%2Fcowboybebop.fandom.com%2Fwiki%2FPembroke_Welsh_Corgi&psig=AOvVaw3qI_c8lyVNLvM0OTllb-aU&ust=1555805441454889",
+        scores: [
+            "3",
+            "5",
+            "1",
+            "3",
+            "2",
+            "4",
+            "2",
+            "4",
+            "3",
+            "4"
+        ]
+    },
+    {
+        name: "Faye Valentine",
+        photo: "https://cdn-images-1.medium.com/max/2400/1*FMDfrKABIsWxd9qIYrUyUQ.jpeg",
+        scores: [
+            "1",
+            "2",
+            "5",
+            "2",
+            "3",
+            "1",
+            "2",
+            "5",
+            "2",
+            "1"
+        ]
+    },
+    {
+        name: "Spike Spiegel",
+        photo: "https://ih1.redbubble.net/image.383258389.9288/ap,550x550,12x12,1,transparent,t.u2.png",
+        scores: [
+            "2",
+            "5",
+            "3",
+            "1",
+            "4",
+            "2",
+            "3",
+            "4",
+            "1",
+            "4"
+        ]
+    },
+    {
+        name: "Jet Black",
+        photo: "https://vignette.wikia.nocookie.net/cowboybebop/images/9/94/Jet_Head.jpg/revision/latest?cb=20130709230234",
+        scores: [
+            "2",
+            "5",
+            "3",
+            "1",
+            "4",
+            "2",
+            "3",
+            "4",
+            "1",
+            "4"
+        ]
+    },
+    {
+        name: "Radical Edward",
+        photo: "https://vignette.wikia.nocookie.net/cowboybebop/images/9/9d/9_EdFlyBebop2.png/revision/latest?cb=20090406042712",
+        scores: [
+            "5",
+            "4",
+            "5",
+            "5",
+            "5",
+            "5",
+            "5",
+            "3",
+            "2",
+            "5"
+        ]
+    },
+]
