+++ app/public/home.html
@@ -20,7 +20,7 @@
       <h3>Answer a few basic questions</h3>
       <h3>Then find the friend with the greatest compatibility!</h3>
       <hr>
-      <a href="#"><button class="btn btn-primary btn-lg"><span class=""></span> Go to Survey</button></a>
+      <a href="/survey"><button class="btn btn-primary btn-lg"><span class=""></span> Go to Survey</button></a>
     </div>
   </div>
 <footer class="footer">
