+++ app/public/survey.html
@@ -134,12 +134,18 @@
         <option value="4">4</option>
         <option value="5">5 (Strongly Agree)</option>
         </select>
+        <br>
+        <br>
+
+        <button type="submit" class="btn btn-primary btn-lg btn-block" id="submit">Submit</button>
+
+        <footer class="footer">
+            <div class="container">
+                <p><a href="#">API Friends List</a> | <a href="#">Github Repository</a></p>
+            </div>
+        </footer>
     </div>
-    <footer class="footer">
-        <div class="container">
-            <p><a href="#">API Friends List</a> | <a href="#">Github Repository</a></p>
-        </div>
-    </footer>
+    
 
 </body>
