+++ app/routing/htmlRoutes.js
@@ -1,48 +1,11 @@
-// Dependencies
-// =============================================================
-var express = require("express");
 var path = require("path");
 
-// Sets up the Express App
-// =============================================================
-var app = express();
-var PORT = 3000;
+module.exports = function (app) {
+    app.get("/survey", function (req, res) {
+        res.sendFile(path.join(__dirname, "../public/survey.html"));
+    });
 
-// Sets up the Express app to handle data parsing
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-
-// Basic route that sends the user first to the AJAX Page
-app.get("/survey", function (req, res) {
-    res.sendFile(path.join(__dirname, "home.html"));
-});
-
-app.get("/add", function (req, res) {
-    res.sendFile(path.join(__dirname, "add.html"));
-});
-
-// Displays all characters
-app.get("/api/characters", function (req, res) {
-    return res.json(characters);
-});
-
-// Displays a single character, or returns false
-app.get("/api/characters/:character", function (req, res) {
-    var chosen = req.params.character;
-
-    console.log(chosen);
-
-    for (var i = 0; i < characters.length; i++) {
-        if (chosen === characters[i].routeName) {
-            return res.json(characters[i]);
-        }
-    }
-
-    return res.json(false);
-});
-
-// Starts the server to begin listening
-// =============================================================
-app.listen(PORT, function () {
-    console.log("App listening on PORT " + PORT);
-});
+    app.get("/", function (req, res) {
+        res.sendFile(path.join(__dirname, "../public/home.html"));
+    });
+}
