+++ app/server.js
@@ -10,6 +10,8 @@ var PORT = process.env.PORT || 3000;
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
+require("./routing/htmlRoutes")(app);
+
 app.listen(PORT, function() {
     console.log("App listening on PORT " + PORT);
   });
